# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация для сервиса справочники",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cities-lang": {
            "get": {
                "tags": [
                    "city-lang"
                ],
                "summary": "List of city lang",
                "description": "Get list of city lang",
                "operationId": "getCitiesLang",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            }
        },
        "/api/city-lang": {
            "post": {
                "tags": [
                    "city-lang"
                ],
                "summary": "Create city lang",
                "description": "Create city lang",
                "operationId": "createCityLang",
                "requestBody": {
                    "description": "Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "english"
                                    },
                                    "lang": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            }
        },
        "/api/city-lang/{id}": {
            "get": {
                "tags": [
                    "city-lang"
                ],
                "summary": "Get city lang",
                "description": "Get city lang",
                "operationId": "getCityLang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "city lang ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            },
            "put": {
                "tags": [
                    "city-lang"
                ],
                "summary": "Update city lang",
                "description": "Update city lang",
                "operationId": "updateCityLang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "city lang ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "lang": {
                                        "type": "string"
                                    },
                                    "city_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            },
            "delete": {
                "tags": [
                    "city-lang"
                ],
                "summary": "Delete city lang",
                "description": "Delete city lang",
                "operationId": "deleteCityLang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "city lang ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            }
        },
...
    },
...
}